apps/hotel-search/requirements.txt
fastapi
uvicorn
redis
python-jose
PyJWT==2.8.0
starlette>=0.27.0